server/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-socketio==5.11.4
pydantic==2.10.6
beautifulsoup4==4.12.3
httpx==0.28.1
